// src/enums.ts
export enum Source {
    IMDB = "IMDB",
    MetaCritics = "metaCritics",
    Unknown = "unknown",
}
export enum Genre {
    Action = "action",
    Adventure = "adventure",
    Animation = "animation",
    Biography = "biography",
    Comedy = "comedy",
    Crime = "crime",
    Documentary = "documentary",
    Drama = "drama",
    Family = "family",
    Fantasy = "fantasy",
    FilmNoir = "filmNoir",
    GameShow = "gameShow",
    History = "history",
    Horror = "horror",
    Music = "music",
    Musical = "musical",
    Mystery = "mystery",
    News = "news",
    RealityTV = "realityTv",
    Romance = "romance",
    SciFi = "sciFi",
    Short = "short",
    Sport = "sport",
    TalkShow = "talkShow",
    Thriller = "thriller",
    War = "war",
    Western = "western",
}

export enum TitleMainType {
    Movie = "movie",
    Series = "series",
    SeriesEpisode = "seriesEpisode",
    TVSpecial = "tvSpecial",
    TVShort = "tvShort",
    TVMovie = "tvMovie",
    Video = "video",
}

export enum Language {
    Persian = "persian",
    English = "english",
}

export enum AwardOutcome {
    Winner = "winner",
    Nominee = "nominee",
}

export enum ImageType {
    Poster = "poster",
    StillFrame = "stillFrame",
    Event = "event",
    VideoImage = "videoImage",
    ProfileImage = "profileImage",
    Other = "other",
}

export enum IMDBPathType {
    Title = "title",
    Name = "name",
}